+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/github/GitHubConnector.java
@@ -20,15 +20,15 @@ import me.mrCookieSlime.Slimefun.api.Slimefun;
 abstract class GitHubConnector {
 
     protected File file;
+    protected String repository;
     protected final GitHubService github;
 
-    public GitHubConnector(GitHubService github) {
+    public GitHubConnector(GitHubService github, String repository) {
         this.github = github;
+        this.repository = repository;
     }
 
     public abstract String getFileName();
-
-    public abstract String getRepository();
 
     public abstract String getURLSuffix();
 
@@ -42,11 +42,11 @@ abstract class GitHubConnector {
         file = new File("plugins/Slimefun/cache/github/" + getFileName() + ".json");
 
         if (github.isLoggingEnabled()) {
-            Slimefun.getLogger().log(Level.INFO, "Retrieving {0}.json from GitHub...", this.getFileName());
+            Slimefun.getLogger().log(Level.INFO, "Retrieving {0}.json from GitHub...", getFileName());
         }
 
         try {
-            URL website = new URL("https://api.github.com/repos/" + this.getRepository() + this.getURLSuffix());
+            URL website = new URL("https://api.github.com/repos/" + repository + getURLSuffix());
 
             URLConnection connection = website.openConnection();
             connection.setConnectTimeout(8000);