+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/weapons/SlimefunBow.java
@@ -3,8 +3,8 @@ package io.github.thebusybiscuit.slimefun4.implementation.items.weapons;
 import org.bukkit.entity.Arrow;
 import org.bukkit.inventory.ItemStack;
 
-import me.mrCookieSlime.Slimefun.Lists.Categories;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
+import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.Objects.handlers.BowShootHandler;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
@@ -19,8 +19,8 @@ import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
  */
 public abstract class SlimefunBow extends SlimefunItem {
 
-    public SlimefunBow(SlimefunItemStack item, ItemStack[] recipe) {
-        super(Categories.WEAPONS, item, RecipeType.MAGIC_WORKBENCH, recipe);
+    public SlimefunBow(Category category, SlimefunItemStack item, ItemStack[] recipe) {
+        super(category, item, RecipeType.MAGIC_WORKBENCH, recipe);
     }
 
     @Override