+++ src/main/java/me/mrCookieSlime/Slimefun/Lists/RecipeType.java
@@ -70,16 +70,6 @@ public class RecipeType implements Keyed {
         this.key = new NamespacedKey(SlimefunPlugin.instance, "null");
     }
 
-    /**
-     * @deprecated Use the constructor with {@link NamespacedKey} instead
-     * @param item
-     *            The {@link ItemStack} to use for this {@link RecipeType}
-     */
-    @Deprecated
-    public RecipeType(ItemStack item) {
-        this(item, "");
-    }
-
     public RecipeType(ItemStack item, String machine) {
         this.item = item;
         this.machine = machine;